refactor(tile): split decode.rs into decode/ submodules + add tests
`decode.rs` was ~400 LOC of mixed concerns: types, the proto include,
geometry-stream decoding, tag/value decoding, gzip detection, and the
top-level `decode()` orchestrator. Splitting clarifies which file
owns which transform and gives the upcoming decoder lane (issue
#69 / future PR) an obvious import target — it just calls
`decode::decode`.

Layout:
- `decode/mod.rs` — public types (`TilePoint`, `Feature`, `TileLayer`,
  `DecodedTile`), the `proto` module, `calculate_bounds`, and the
  top-level `decode()` entry.
- `decode/geometry.rs` — `zigzag` + `decode_geometry` (the MVT
  command-stream parser; this is where the #102 bounds-check fix
  lives).
- `decode/tags.rs` — `decode_tags_into` + `proto_value_to_pv`.
- `decode/decompress.rs` — `maybe_decompress` (gzip sniff + inflate).

Tests added while files were small and focused (the user-prompted
companion to the split):
- `decompress`: gzip round-trip, empty input, corrupt-magic-bytes
  guard.
- `tags`: paired-index happy path, out-of-range key/value, `None`-
  in-pool, dangling odd index, plus per-branch coverage of
  `proto_value_to_pv` (string-precedence, bool-precedence, all
  numeric collapse paths, empty value).
- `mod`: extra `calculate_bounds` cases (empty input, single point,
  multi-ring union) and `decode` smoke tests for empty + garbage
  protobuf.

Test count: 260 → 277 (+17 new).

Behavior change: zero. Visibility: helper fns are `pub(super)` so the
parent module can call them; nothing in the file is newly exported.
Doc updates in CLAUDE.md / README to point at `decode/` instead of
`decode.rs`.

Third (and last) of the mechanical prereqs before the substantive
`TileFetcher` / `FetchLane` redesign.
